- patch from Stefan Farfeleder...
here is a patch that enables me to run bbkeys 0.9.0 on FreeBSD
6-current.  You forgot to set the members _keybindings and
config_check_timer in the constructor to 0, which caused the
initialisation code to be skipped (KeyClient.cpp:194 and
KeyClient.cpp:152).  If you wonder why this didn't happen to you,
FreeBSD's malloc() fills allocated memory with 0xd0 to catch such
errors.